% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioassay_survival_to_resistance.R
\name{bioassay_survival_to_resistance}
\alias{bioassay_survival_to_resistance}
\title{Convert mosquito bioassay survival to mean population  resistance intensity}
\usage{
bioassay_survival_to_resistance(
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  bioassay.survival = 0.1,
  estimate.precision = 0.01,
  sd.population.resistance = 10,
  nsim = 1000,
  minimum.resistance.value = 0,
  maximum.resistance.value = 25000
)
}
\arguments{
\item{maximum.bioassay.survival.proportion}{Should be set as 1.}

\item{michaelis.menten.slope}{Should be set as 1}

\item{half.population.bioassay.survival.resistance}{This is calculated using the calculate_half_population_resistance function}

\item{bioassay.survival}{The survival that occured in the bioassay, as a proportion(values must be between 0 and 1). Where 1=all survived, 0=all died}

\item{estimate.precision}{How precise your estimate of insecticide restistance intensity should be. Recommend values between 0.01 to 0.001}

\item{sd.population.resistance}{How much variation in the population resistance.}

\item{nsim}{How many replications of the rnorm function are conducted. Recommended value is 1000.}

\item{minimum.resistance.value}{Recommend setting as 0.}

\item{maximum.resistance.value}{This will depend on the half survival scale, but 10000 would be a good start.}
}
\value{
test.population.resistance This is the mean insecticide resistance intensity of the population to the respected insecticide.
}
\description{
Converts the proportion of individuals surviving in a bioassay (eg. CDC Bottle or WHO Cylinder) into
an insecticide resistance intensity score.
}
