% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_half_population_survival.R
\name{calculate_half_population_survival}
\alias{calculate_half_population_survival}
\title{Calculate the 50\% survival for a given resistance intensity}
\usage{
calculate_half_population_survival(
  desired.resistance,
  desired.survival.proportion,
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  estimate.precision = 0.01,
  sd.population.resistance,
  nsim = 1000,
  minimum.resistance.value = 0,
  maximum.resistance.value = 25000
)
}
\arguments{
\item{desired.resistance}{The value you want a survival value to correspond to}

\item{desired.survival.proportion}{The survival proportion you want your desired.resistance to have. Values must be between 0 and 1}

\item{maximum.bioassay.survival.proportion}{Must be set at 1}

\item{michaelis.menten.slope}{Must be set at 1}

\item{estimate.precision}{How precise your estimate of insecticide restistance intensity should be. Recommend values between 0.01 to 0.001}

\item{sd.population.resistance}{How much variation in the population resistance. Recommend setting as zero for scale setting}

\item{nsim}{Number of replications of the rnorm function. Recommended value = 1000}

\item{minimum.resistance.value}{Recommend setting to 0. Must be lower than half survival resistance.}

\item{maximum.resistance.value}{Depends on your scale. Recommend setting arbitrarily high (10000). Must be higher than half resistance survival}
}
\value{
a value for the half.population.survival.value which is the half.population.bioassay.survival.resistance for the scale.
}
\description{
A function that enables calculation of a  50\% survival threshold, based on desired
resistance value and its desired survival value. For example if you want to base the scale on
a 20\% survival having a resistance intensity of 1000.
}
