% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_insecticide_withdrawal.R
\name{check_insecticide_withdrawal}
\alias{check_insecticide_withdrawal}
\title{Check if insecticides should be withdrawn from the arsenal based on the resistance intensity.}
\usage{
check_insecticide_withdrawal(
  insecticide,
  current.generation,
  withdrawal.threshold,
  simulation.array
)
}
\arguments{
\item{insecticide}{= The insecticide to which the resistance intensity is to be checked against}

\item{current.generation}{= The generation where the run_simulation() is up to.}

\item{withdrawal.threshold}{= The resistance intensity at which an insecticide should no longer be used and therefore withdrawn.}

\item{simulation.array}{= The array which holds the run_simulation results.}
}
\description{
Check if insecticides should be withdrawn from the arsenal based on the resistance intensity.
}
