% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_next_insecticide.R
\name{choose_next_insecticide}
\alias{choose_next_insecticide}
\title{Choose the next insecticide for deployment.}
\usage{
choose_next_insecticide(
  previous.insecticide,
  available.insecticides,
  number.of.insecticides
)
}
\arguments{
\item{previous.insecticide}{= The insecticide that was last deployed.}

\item{available.insecticides}{= A vector containing the available insecticides.}

\item{number.of.insecticides}{= The total number of insecticides (available and withdrawn) that are in the simulation.}
}
\value{
candidate.insecticide = a single integer which represents the insecticide which is to be deployed.
}
\description{
The function selects the next available insecticide for deployment. It does not choose based on which would be the optimal
insecticide (eg. the insecticide with the lowest resistance intensity), but which is next based on the numerical sequence of
the insecticides.
}
