% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_starting_array.R
\name{create_starting_array}
\alias{create_starting_array}
\title{Create the array for holding the simulation.}
\usage{
create_starting_array(n.insecticides, maximum.generations)
}
\arguments{
\item{n.insecticides}{the number of different insecticides in the armoury.}

\item{maximum.generation}{the maximum number of generations the model would be able to run for.}
}
\value{
simulation.data an array which will hold the data for simulation
}
\description{
Returns an array: site, insecticide, generation.
The site will be refugia or treatment. And will hold the corresponding resistance intensity values. The insecticide is the
insecticide to which the resistance intensity corresponds to. The generation is the timepoint in the model to which the intensity
corresponds to.
}
