% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_insecticide.R
\name{deploy_insecticide}
\alias{deploy_insecticide}
\title{Deploys selected insecticide for the for a specified timeframe (in generations):}
\usage{
deploy_insecticide(
  insecticide.to.deploy,
  deployment.frequency,
  deployment.vector
)
}
\arguments{
\item{insecticide.to.deploy}{= The value of the insecticide to be deployed}

\item{deployment.frequency}{= The number of generations between each insecticide deployment \link{5 generations is 6months, 10 generations is 12months}}

\item{deployment.vector}{= The vector which holds to sequence of insecticides that have been deployed,}
}
\description{
Deploys selected insecticide for the for a specified timeframe (in generations):
}
