% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_of_fitness_cost.R
\name{effect_of_fitness_cost}
\alias{effect_of_fitness_cost}
\title{Implements the fitness cost associated with having insecticide resistance.}
\usage{
effect_of_fitness_cost(
  resistance.cost = 0,
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9
)
}
\arguments{
\item{resistance.cost}{Costs associated with insecticide resistance. Values recommended between 0.01 to 0.2}

\item{nsim}{number of simulations of the runif functions}

\item{minimum.insecticide.resistance.heritability}{minimum heritability of IR (default is 0.05)}

\item{maximum.insecticide.resistance.heritability}{maximum heritability of IR (default is 0.30)}

\item{minimum.male.insecticide.exposure}{Proportion of males exposed as a proportion of females (default is 0)}

\item{maximum.male.insecticide.exposure}{Proportion of males exposed as a proportion of females (default is 1)}

\item{minimum.female.insecticide.exposure}{Proportion of females exposed at intervention site (default = 0.4)}

\item{maximum.female.insecticide.exposure}{Proportion of females exposed at intervention site (default = 0.9)}
}
\value{
fitness.cost  in population resistance intensity. Will be a negative number.
}
\description{
Implements the fitness cost associated with having insecticide resistance.
}
