% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulation_dataframe.R
\name{get_simulation_dataframe}
\alias{get_simulation_dataframe}
\title{Convert the array from the simulation into a dataframe}
\usage{
get_simulation_dataframe(
  simulation.array,
  maximum.generations,
  number.of.insecticides
)
}
\arguments{
\item{simulation.array}{= the array which holds the simulation data called from run_simulation_intervention.}

\item{maximum.generations}{= The number of generations the simulation was asked to run for (maximum.generations)}

\item{number.of.insecticides}{= the number of insectcides that are tracked in the simulation}
}
\value{
final.df A dataframe that has 5 columns, insecticide.tracked, resistance.intensity, site,
time.in.generations and insecticide.deployed.
}
\description{
Input the saved simulation array into this function to return a dataframe holding the information of regarding the
simulation. The dataframe has 5 columns. insecticide.tracked is the insecticide to which the resistance.intensity corresponds to
at the site (either refugia or treatment), at the timepoint time.in.generations. When the corresponding insecticide.deployed is the
insecticide that is deployed in the treatment site at at that generation.
}
