% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bioassay_survival_to_resistance.R,
%   R/plot_michaelis_menten_slope_resistance.R,
%   R/plot_michaelis_menten_slope_survival.R,
%   R/plot_resistance_to_bioassay_survival.R,
%   R/table_michaelis_menten_slope_survival.R,
%   R/table_resistance_from_survival_and_sd.R
\name{\%>\%}
\alias{\%>\%}
\alias{plot_bioassay_survival_to_resistance}
\alias{plot_michaelis_menten_slope_resistance}
\alias{plot_michaelis_menten_slope_survival}
\alias{plot_resistance_to_bioassay_survival}
\alias{table_michaelis_menten_slope_survival}
\alias{table_resistance_from_survival_and_sd}
\title{Plot the relationship between survival and resistance calculated from bioassay survival}
\usage{
plot_bioassay_survival_to_resistance(
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  estimate.precision = 0.01,
  sd.population.resistance = 10,
  nsim = 1000,
  minimum.resistance.value = 0,
  maximum.resistance.value = 25000,
  minimum.bioassay.survival = 0,
  maximum.bioassay.survival = 1,
  divisions = 0.01
)

plot_michaelis_menten_slope_resistance(
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope.values,
  half.population.bioassay.survival.resistance = 900,
  bioassay.survival = 0.5,
  estimate.precision = 0.01,
  sd.population.resistance.values,
  nsim = 1000,
  minimum.resistance.value = 0,
  maximum.resistance.value = 25000
)

plot_michaelis_menten_slope_survival(
  michaelis.menten.slope.values,
  sd.population.resistance.values,
  maximum.bioassay.survival = 1,
  half.population.bioassay.survival.resistance = 900,
  mean.population.resistance,
  nsim
)

plot_resistance_to_bioassay_survival(
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  nsim = 1000,
  minimum.resistance = 0,
  maximum.resistance = 25000,
  sd.population.resistance = 10
)

table_michaelis_menten_slope_survival(
  michaelis.menten.slope.values,
  sd.population.resistance.values,
  maximum.bioassay.survival.proportion = 1,
  half.population.bioassay.survival.resistance = 900,
  mean.population.resistance = 900,
  nsim = 1000
)

table_resistance_from_survival_and_sd(
  half.population.bioassay.survival.resistance = 900,
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  bioassay.survival.values,
  sd.population.values,
  estimate.precision = 0.01,
  nsim = 1000,
  minimum.resistance.value = 0,
  maximum.resistance.value = 25000
)
}
\arguments{
\item{maximum.bioassay.survival.proportion}{Should be set as 1}

\item{michaelis.menten.slope}{Should be set as 1}

\item{half.population.bioassay.survival.resistance}{As calculated using calculate_half_population_survival}

\item{estimate.precision}{ow precise your estimate of insecticide restistance intensity should be. Recommend values between 0.01 to 0.001}

\item{sd.population.resistance}{How much variation in the population resistance.}

\item{nsim}{How many replications of the rnorm function are conducted. Recommended value is 1000.}

\item{minimum.resistance.value}{Should be set as 0.}

\item{maximum.resistance.value}{Value required would depend on resistance intensity scale. Recommend starting with 10000}

\item{minimum.bioassay.survival}{As a proportion, should be set as 0}

\item{maximum.bioassay.survival}{Must be set at 1.}

\item{divisions}{The resolution of the final plotted graph}

\item{michaelis.menten.slope.values}{A vector of values, including the value 1. Recommend a sequence from 0 to 2.}

\item{bioassay.survival}{The survival that occured in the bioassay, as a proportion(values must be between 0 and 1). Where 1=all survived, 0=all died}

\item{sd.population.resistance.values}{How much variation in the population resistance.}

\item{mean.population.resistance}{The resistance intensity of the population}

\item{minimum.resistance}{Recommend setting as 0.}

\item{maximum.resistance}{This will depend on the half survival scale, but 10000 would be a good start.}

\item{bioassay.survival.values}{A vector containing survival values with repeats. Same length as sd.population.values.}

\item{sd.population.values}{A vector containing repeated standard deviation values. Same length as bioassay.survival values}

\item{michaelis.menton.slope.values}{A vector of values, including the value 1. Recommend a sequence from 0 to 2.}
}
\value{
A ggplot of the relationship between resistance (x axis) and survival (y axis).

A ggplot of the relationship between resistance (x axis) and survival (y axis).
}
\description{
Plot the relationship between survival and resistance calculated from bioassay survival

This function demonstrates the impact (by plotting) of what happens when you change michaelis.menten.slope in the bioassay_survival_to_resistance function,
and demonstrates how the function only performs correctly when the michaelis.menten.slope = 1.

Plot the impact of changing the michaelis.menten.slope parameter in the resistance_to_bioassay_survival function.
Demonstrates the equations only work correctly when michaelis.menten.slope = 1.

Plot the relationship between resistance and survival calculated from insecticide resistance

Create a table of the impact of michaelis.menten.slope values

Function to generate a table of resistant values based on bioassay survival and standard deviation
}
