% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insecticide_deployed_migration.R
\name{insecticide_deployed_migration}
\alias{insecticide_deployed_migration}
\title{The effect of dispersal and resistance costs when the insecticide is deployed.}
\usage{
insecticide_deployed_migration(
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9,
  resistance.cost = 0,
  initial.resistance.intensity,
  min.intervention.coverage = 0.1,
  max.intervention.coverage = 0.9,
  initial.refugia.resistance,
  min.dispersal.rate = 0.1,
  max.dispersal.rate = 0.9
)
}
\arguments{
\item{exposure.scaling.factor}{= 10,}

\item{nsim}{= 1000,}

\item{minimum.insecticide.resistance.heritability}{= 0.05,}

\item{maximum.insecticide.resistance.heritability}{= 0.30,}

\item{minimum.male.insecticide.exposure}{= 0,}

\item{maximum.male.insecticide.exposure}{= 1,}

\item{minimum.female.insecticide.exposure}{= 0.4,}

\item{maximum.female.insecticide.exposure}{= 0.9,}

\item{min.intervention.coverage}{= 0.1,}

\item{max.intervention.coverage}{= 0.9,}

\item{min.dispersal.rate}{= 0.1}

\item{max.dispersal.rate}{= 0.9}
}
\value{
resistance.intensity.migration A vector of length nsim
}
\description{
This function returns the population mean resistance for the treated location, when
insecticide is deployed, after migration and selection costs.
}
