% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insecticide_deployed_selection_cost.R
\name{insecticide_deployed_selection_cost}
\alias{insecticide_deployed_selection_cost}
\title{Implement  the selection costs when the insecticide is present in the treated site.}
\usage{
insecticide_deployed_selection_cost(
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9,
  resistance.cost = 0.1,
  initial.resistance.intensity = 0
)
}
\arguments{
\item{fitness.cost}{}
}
\value{
resistance.intensity.values
}
\description{
Implements the selection costs when insecticide is present in the treated site Does not include dispersal effect.
Gives the resistance intensity in the next generation afer selection costs and before migration in the treated site.
}
