% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insecticide_not_deployed_migration.R
\name{insecticide_not_deployed_migration}
\alias{insecticide_not_deployed_migration}
\title{Effect of migration when the corresponding insecticide is not deployed.}
\usage{
insecticide_not_deployed_migration(
  initial.resistance.intensity,
  resistance.cost,
  initial.refugia.resistance,
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9,
  min.intervention.coverage = 0.1,
  max.intervention.coverage = 0.9,
  min.dispersal.rate = 0.1,
  max.dispersal.rate = 0.9
)
}
\arguments{
\item{initial.resistance.intensity}{This is the resistance intensity in the treatment site}

\item{initial.refugia.resistance}{This is the resistance intensity in the refugia}

\item{exposure.scaling.factor}{= 10,}

\item{nsim}{= 1000,}

\item{minimum.insecticide.resistance.heritability}{= 0.05,}

\item{maximum.insecticide.resistance.heritability}{= 0.30,}

\item{minimum.male.insecticide.exposure}{= 0,}

\item{maximum.male.insecticide.exposure}{= 1,}

\item{minimum.female.insecticide.exposure}{= 0.4,}

\item{maximum.female.insecticide.exposure}{= 0.9,}

\item{min.intervention.coverage}{= 0.1,}

\item{max.intervention.coverage}{= 0.9,}

\item{min.dispersal.rate}{= 0.1,}

\item{max.dispersal.rate}{= 0.9}
}
\description{
Effect of migration when the corresponding insecticide is not deployed.
}
