% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insecticide_not_deployed_selection_cost.R
\name{insecticide_not_deployed_selection_cost}
\alias{insecticide_not_deployed_selection_cost}
\title{Track the resistance intensity when the insecticide is not deployed}
\usage{
insecticide_not_deployed_selection_cost(
  initial.resistance.intensity,
  resistance.cost,
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9
)
}
\arguments{
\item{exposure.scaling.factor}{= 10,}

\item{nsim}{= 1000,}

\item{minimum.insecticide.resistance.heritability}{= 0.05,}

\item{maximum.insecticide.resistance.heritability}{= 0.30,}

\item{minimum.male.insecticide.exposure}{= 0,}

\item{maximum.male.insecticide.exposure}{= 1,}

\item{minimum.female.insecticide.exposure}{= 0.4,}

\item{maximum.female.insecticide.exposure}{= 0.9}
}
\value{
track.resistance.intensity
}
\description{
Tracks the insecticide resistance intensity when the corresponding insecticide is
not deployed in the treatment site. But before the effective of mosquito migration has taken place.
}
