% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irm_strategy_rotation.R
\name{irm_strategy_rotation}
\alias{irm_strategy_rotation}
\title{Implement the rotation insectidicide resistance management strategy.}
\usage{
irm_strategy_rotation(
  number.of.insecticides,
  current.generation,
  withdrawal.threshold,
  return.threshold,
  simulation.array,
  available.vector,
  withdrawn.vector,
  current.insecticide,
  deployment.frequency,
  deployment.vector
)
}
\arguments{
\item{number.of.insecticides}{= The total number of insecticides included in the simulation.}

\item{current.generation}{= The generation where the simulation is up to.}

\item{withdrawal.threshold}{= The resistance intensity that leads to an insecticide being withdrawn}

\item{return.threshold}{= The resistance intensity that allows an insecticide to be returned to deployment}

\item{simulation.array}{= The array which holds the simulation results}

\item{available.vector}{= A vector containing the insecticides which can be used for deployment}

\item{withdrawn.vector}{= A vector the insecticides which are withdrawn from deployment}

\item{current.insecticide}{= The insecticide that is currently in deployment}

\item{deployment.frequency}{= The number of mosquito generations between each insecticide deployment check}

\item{deployment.vector}{= A vector containing the sequence of insecticides that have been deployed.}
}
\value{
list(available.to.deploy, unavailable.to.deploy, deployment.vector.updated)
}
\description{
The implementation of the rotation strategy means that a different insecticide must be deployed
for each deployment interval. For example if insecticide 1 was deployed in the previous deployment it could not
be deployed in the next deployment. A list of three updated vectors is returned:
available.vector[\link{1}], withdrawn.vector[\link{2}] and deployed.vector[\link{3}]
}
