% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irm_strategy_sequence.R
\name{irm_strategy_sequence}
\alias{irm_strategy_sequence}
\title{Implement the deployment of the sequence insecticide resistance management strategy.}
\usage{
irm_strategy_sequence(
  number.of.insecticides,
  current.generation,
  withdrawal.threshold,
  return.threshold,
  simulation.array,
  available.vector,
  withdrawn.vector,
  current.insecticide,
  deployment.frequency,
  deployment.vector
)
}
\arguments{
\item{number.of.insecticides}{= The total number of insecticides included in the simulation.}

\item{current.generation}{= The generation where the simulation is up to.}

\item{withdrawal.threshold}{= The resistance intensity that leads to an insecticide being withdrawn}

\item{return.threshold}{= The resistance intensity that allows an insecticide to be returned to deployment}

\item{simulation.array}{= The array which holds the simulation results}

\item{available.vector}{= A vector containing the insecticides which can be used for deployment}

\item{withdrawn.vector}{= A vector the insecticides which are withdrawn from deployment}

\item{current.insecticide}{= The insecticide that is currently in deployment}

\item{deployment.frequency}{= The number of mosquito generations between each insecticide deployment check}

\item{deployment.vector}{= A vector containing the sequence of insecticides that have been deployed.}
}
\value{
list(available.to.deploy, unavailable.to.deploy, deployment.vector.updated)
}
\description{
Implementation of the sequence IRM strategy allows for an insecticide to be continously used
until that insecticide reaches the threshold for withdrawal. After this the next insecticide is deployed.
This function returns a list of the updated vectors of the available.vector[\link{1}], withdrawn.vector[\link{2}]
and the deployed.vector[\link{3}]
}
