% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_treatment_to_refugia.R
\name{migration_treatment_to_refugia}
\alias{migration_treatment_to_refugia}
\title{Calculate the migration effect from the treatment area to the refugia.}
\usage{
migration_treatment_to_refugia(
  nsim = 1000,
  min.intervention.coverage = 0.1,
  max.intervention.coverage = 0.9,
  min.dispersal.rate = 0.1,
  max.dispersal.rate = 0.9
)
}
\arguments{
\item{nsim}{Number of interations of the runif function}

\item{min.intervention.coverage}{Minimum coverage of uniform distribution of insecticide intervention, set to 0.1}

\item{max.intervention.coverage}{Minimum coverage of uniform distribution of insecticide intervention, set to 0.9}

\item{min.dispersal.rate}{Minimum disperal proportion of uniform distribution, Set to 0.1}

\item{max.dispersal.rate}{Maximum dispersal propotion of unifrom distribution, set to 0.9}
}
\value{
exchange.treatment.to.refugia
}
\description{
Calculate the migration effect from the treatment area to the refugia.
}
