% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_response_to_insecticide_selection.R
\name{plot_log_response_to_insecticide_selection}
\alias{plot_log_response_to_insecticide_selection}
\title{Function to plot a histogram of the response_to_insecticide_selection output.
Plot to give years to 10\% Survival (assuming 10 generations per year)}
\usage{
plot_log_response_to_insecticide_selection(response.values)
}
\arguments{
\item{response.values}{As obtained from the response_to_insecticide_selection function}
}
\description{
Function to plot a histogram of the response_to_insecticide_selection output.
Plot to give years to 10\% Survival (assuming 10 generations per year)
}
