% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refugia_migration_effect.R
\name{refugia_migration_effect}
\alias{refugia_migration_effect}
\title{The effect of migration on the resistance intensity in the refugia}
\usage{
refugia_migration_effect(
  initial.refugia.resistance,
  initial.resistance.intensity,
  resistance.cost,
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9,
  min.intervention.coverage = 0.1,
  max.intervention.coverage = 0.9,
  min.dispersal.rate = 0.1,
  max.dispersal.rate = 0.9
)
}
\arguments{
\item{initial.refugia.resistance}{The current IR intensity in the refugia.}

\item{resistance.cost}{The cost of having IR}

\item{exposure.scaling.factor}{factor converting exposure into selection differential (default is 10)}

\item{nsim}{number of simulations of the runif functions}

\item{minimum.insecticide.resistance.heritability}{minimum heritability of IR (default is 0.05)}

\item{maximum.insecticide.resistance.heritability}{maximum heritability of IR (default is 0.30)}

\item{minimum.male.insecticide.exposure}{Proportion of males exposed as a proportion of females (default is 0)}

\item{maximum.male.insecticide.exposure}{Proportion of males exposed as a proportion of females (default is 1)}

\item{minimum.female.insecticide.exposure}{Proportion of females exposed at intervention site (default = 0.4)}

\item{maximum.female.insecticide.exposure}{Proportion of females exposed at intervention site (default = 0.9)}

\item{min.intervention.coverage}{Proportion of the insect population in the treated intervention area (default is 0.1)}

\item{max.intervention.coverage}{Proportion of the insect population in the treated intervention area (default is 0.9)}

\item{min.dispersal.rate}{Proportion of population exchanged between the treatment and refugia sites per generation (default 0.1)}

\item{max.dispersal.rate}{Proportion of the population exchanged between the treatment and refugia sites per generation}
}
\value{
track.refugia.resistance A vector length nsim of resistance intensity values in the refugia
}
\description{
The effect of migration on the resistance intensity in the refugia
}
