% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refugia_selection_costs.R
\name{refugia_selection_costs}
\alias{refugia_selection_costs}
\title{Implement the selection costs in the refugia}
\usage{
refugia_selection_costs(
  initial.refugia.resistance,
  resistance.cost,
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9
)
}
\arguments{
\item{exposure.scaling.factor}{= 10,}

\item{nsim}{= 1000,}

\item{minimum.insecticide.resistance.heritability}{= 0.05,}

\item{maximum.insecticide.resistance.heritability}{= 0.30,}

\item{minimum.male.insecticide.exposure}{= 0,}

\item{maximum.male.insecticide.exposure}{= 1,}

\item{minimum.female.insecticide.exposure}{= 0.4,}

\item{maximum.female.insecticide.exposure}{= 0.9}
}
\value{
track.refugia.resistance is the resistance intensity in the refugia (a vector of length nsim).
}
\description{
Implement the selection costs in the refugia
}
