% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance_to_bioassay_survival.R
\name{resistance_to_bioassay_survival}
\alias{resistance_to_bioassay_survival}
\title{Convert insecticide resistance intensity to bioassay survival proportion}
\usage{
resistance_to_bioassay_survival(
  maximum.bioassay.survival.proportion = 1,
  mean.population.resistance = 900,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  sd.population.resistance = 10,
  nsim = 1000
)
}
\arguments{
\item{maximum.bioassay.survival.proportion}{Should be set as 1.}

\item{mean.population.resistance}{The mean resistance intensity of the population.}

\item{michaelis.menten.slope}{Should be set as 1}

\item{half.population.bioassay.survival.resistance}{This is calculated using the calculate_half_population_resistance function}

\item{sd.population.resistance}{How much variation in the population resistance.}

\item{nsim}{How many replications of the rnorm function are conducted. Recommended value is 1000.}
}
\value{
mean(bioassay.survival.proportion) This is the mean bioassay survival for the population resistance intensity.
}
\description{
Convert insecticide resistance intensity to bioassay survival proportion
}
