% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_and_withdrawal_of_insecticides.R
\name{return_and_withdrawal_of_insecticides}
\alias{return_and_withdrawal_of_insecticides}
\title{Implement both the withdrawal and return of insecticides from the arsenal.}
\usage{
return_and_withdrawal_of_insecticides(
  number.of.insecticides,
  current.generation,
  withdrawal.threshold,
  return.threshold,
  simulation.array,
  available.vector,
  withdrawn.vector
)
}
\arguments{
\item{number.of.insecticides}{= The total number of insecticides in the simulation}

\item{current.generation}{= The current generation in the simualation}

\item{withdrawal.threshold}{= The resistance intensity threshold for withdrawing an insecticide from the arsenal}

\item{return.threshold}{= The resistance intensity threshold for returning an insecticide to the arsenal}

\item{simulation.array}{= The array holding the simulation results}

\item{available.vector}{= A vector containing the previously available insecticides}

\item{withdrawn.vector}{= A vector containing the previously withdrawn insecticides}
}
\value{
list(available.vector.r2, withdrawn.vector.r2) returns a list of the the updated
available[\link{1}] and withdrawn[\link{2}] insecticides.
}
\description{
This function determines both which insecticides are available for deployment and those that
need to be withdrawn from the arsenal. Currently this function is designed around only a single insecticide
deployment, and may need modifying to account for insecticide mixtures.
}
