% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_insecticide_to_arsenal.R
\name{return_insecticide_to_arsenal}
\alias{return_insecticide_to_arsenal}
\title{Transfer a single insecticide from being withdrawn to being available}
\usage{
return_insecticide_to_arsenal(
  withdrawn.vector,
  available.vector,
  insecticide.to.return
)
}
\arguments{
\item{withdrawn.vector}{= A vector containing the insecticides with have been withdrawn from the arsenal.}

\item{available.vector}{= A vector containing the insecticides which are available for deployment.}

\item{insecticide.to.return}{= The value of the insecticide to be returned to the arsenal.}
}
\value{
list(available.vector, withdrawn.vector) : [\link{1}]available vector and [\link{2}]withdrawn vector
}
\description{
A function that returns an insecticide back to the arsenal having been withdrawb.
This function updates the withdrawn and available insecticide vectors for the single insecticide. This is all passed in
the simulations to the return_and_withdrawal_of_insecticides function to perform this task on all the insecicides in the
simulation.Returns a list: [\link{1}]available vector and [\link{2}]withdrawn vector
}
