% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_intervention.R
\name{run_simulation_intervention}
\alias{run_simulation_intervention}
\title{Run the insecticide resistance management simulation for sequences and rotations.}
\usage{
run_simulation_intervention(
  number.of.insecticides = 2,
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9,
  resistance.cost = 0,
  starting.treatment.site.intensity = 0,
  starting.refugia.intensity = 0,
  min.intervention.coverage = 0.1,
  max.intervention.coverage = 0.9,
  min.dispersal.rate = 0.1,
  max.dispersal.rate = 0.9,
  maximum.generations = 500,
  irm.strategy = "sequence",
  half.population.bioassay.survival.resistance = 900,
  withdrawal.threshold.value = 0.1,
  return.threshold.value = 0.05,
  deployment.frequency = 10,
  maximum.resistance.value = 25000
)
}
\arguments{
\item{number.of.insecticides}{how many insecticides are available in the arsenal.}

\item{exposure.scaling.factor}{= 10,}

\item{nsim}{= 1000,}

\item{minimum.insecticide.resistance.heritability}{= 0.05,}

\item{maximum.insecticide.resistance.heritability}{= 0.30,}

\item{minimum.male.insecticide.exposure}{= 0,}

\item{maximum.male.insecticide.exposure}{= 1,}

\item{minimum.female.insecticide.exposure}{= 0.4,}

\item{maximum.female.insecticide.exposure}{= 0.9,}

\item{resistance.cost}{= default set to 0}

\item{min.intervention.coverage}{= 0.1,}

\item{max.intervention.coverage}{= 0.9,}

\item{min.dispersal.rate}{= 0.1,}

\item{max.dispersal.rate}{= 0.9}

\item{irm.strategy}{To be able to set the resistance management strategy (rotation, sequence)}

\item{maximum.resistance.value}{It is recommended that this be set as approximately 20*half.population.bioassay.survival.resistance

Currently this function runs.}
}
\description{
This is the main wrapper function that implements the running of the insecticide resistance management
simulations. Currently the simulations allows for the comparison of sequence and rotation strategies. At the moment,
each insecticide acts completely independently; such that there is no cross resistance and cross selection.
For the "sequence" irm.strategy a single insecticide is continually deployed until it reaches the threshold for withdrawal,
at this point the next insecticide is deployed (at the next deployment opportunity). For the "rotation" irm.strategy, the
insecticide is changed at each deployment interval. It is therefore not recommend to compare rotations and sequences when
the number.of.insecticides = 1; as this will mean that the rotation strategy lasts a single deployment duration.
}
