% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_rotation.R
\name{run_simulation_rotation}
\alias{run_simulation_rotation}
\title{This function runs the model in the absence of dispersal, but with selection costs.}
\usage{
run_simulation_rotation(
  number.of.insecticides = 2,
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9,
  resistance.cost = 0,
  starting.treatment.site.intensity = 0,
  starting.refugia.intensity = 0,
  min.intervention.coverage = 0.1,
  max.intervention.coverage = 0.9,
  min.dispersal.rate = 0.1,
  max.dispersal.rate = 0.9,
  maximum.generations = 500,
  irm.strategy = sequence,
  half.population.bioassay.survival.resistance = 900,
  withdrawal.threshold.value = 0.1,
  return.threshold.value = 0.05,
  deployment.frequency = 10,
  maximum.resistance.value = 25000
)
}
\arguments{
\item{number.of.insecticides}{how many insecticides are available in the arsenal.}

\item{exposure.scaling.factor}{= 10,}

\item{nsim}{= 1000,}

\item{minimum.insecticide.resistance.heritability}{= 0.05,}

\item{maximum.insecticide.resistance.heritability}{= 0.30,}

\item{minimum.male.insecticide.exposure}{= 0,}

\item{maximum.male.insecticide.exposure}{= 1,}

\item{minimum.female.insecticide.exposure}{= 0.4,}

\item{maximum.female.insecticide.exposure}{= 0.9,}

\item{resistance.cost}{= default set to 0}

\item{min.intervention.coverage}{= 0.1,}

\item{max.intervention.coverage}{= 0.9,}

\item{min.dispersal.rate}{= 0.1,}

\item{max.dispersal.rate}{= 0.9}

\item{irm.strategy}{To be able to set the resistance management strategy (rotation, sequence)

Currently this function runs.}
}
\description{
This function runs the model in the absence of dispersal, but with selection costs.
}
