% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_to_insecticide_selection.R
\name{runif}
\alias{runif}
\alias{response_to_insecticide_selection}
\title{Convert insecticide selection into selection response, using the Breeder's Equation}
\usage{
response_to_insecticide_selection(
  exposure.scaling.factor = 10,
  nsim = 1000,
  minimum.insecticide.resistance.heritability = 0.05,
  maximum.insecticide.resistance.heritability = 0.3,
  minimum.male.insecticide.exposure = 0,
  maximum.male.insecticide.exposure = 1,
  minimum.female.insecticide.exposure = 0.4,
  maximum.female.insecticide.exposure = 0.9
)
}
\arguments{
\item{exposure.scaling.factor}{This factor converts exposure to selection differential. This requires empirically setting.}

\item{nsim}{number of simulations of the runif functions.}

\item{minimum.insecticide.resistance.heritability}{The minimum heritability of insecticide resistance (set at 0.05)}

\item{maximum.insecticide.resistance.heritability}{The maximum heritability of insecticide resistance (set at 0.3)}

\item{minimum.male.insecticide.exposure}{Proportion of males exposed to insecticide as a proportion of females. Set at 0.}

\item{maximum.male.insecticide.exposure}{Proportion of males exposed to insecticide as a proportion of females. Set at 1.}

\item{minimum.female.insecticide.exposure}{Mininum proportion of females exposed to insecticide (set at 0.4)}

\item{maximum.female.insecticide.exposure}{Maximum proportion of females exposed to insecticide (set at 0.9)}
}
\value{
resistance.selection.response
}
\description{
Convert insecticide selection into selection response, using the Breeder's Equation
}
