% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_number_of_insecticides.R
\name{set_number_of_insecticides}
\alias{set_number_of_insecticides}
\title{This function sets the number of insecticides to be used in the simulation}
\usage{
set_number_of_insecticides(total.insecticides = 3)
}
\arguments{
\item{total.insecticides}{This is the total number of insecticides available for deployment}
}
\description{
This function sets the number of insecticides to be used in the simulation
}
