% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withdraw_insecticide_from_arsenal.R
\name{withdraw_insecticide_from_arsenal}
\alias{withdraw_insecticide_from_arsenal}
\title{Remove an insecticide from the arsenal. This function updates the withdrawn and available insecticide
vectors. Returns a list: [\link{1}]available vector and [\link{2}]withdrawn vector.}
\usage{
withdraw_insecticide_from_arsenal(
  withdrawn.vector,
  available.vector,
  insecticide.to.withdraw
)
}
\arguments{
\item{withdrawn.vector}{= A vector containing the insecticides with have been withdrawn from the arsenal.}

\item{available.vector}{= A vector containing the insecticides which are available for deployment.}

\item{insecticide.to.withdraw}{= The value of the insecticide to be withdrawn}
}
\description{
Remove an insecticide from the arsenal. This function updates the withdrawn and available insecticide
vectors. Returns a list: [\link{1}]available vector and [\link{2}]withdrawn vector.
}
